requirement.txt
django
django-allauth
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
Pillow
stripe
python-decouple